Add getCurrentDateTime function to LineChart component
