+++ demo/components/Charts/LineChart.tsx
@@ -5,6 +5,7 @@ import { Chart } from 'primereact/chart';
 import jsPDF from 'jspdf';
 import { Button } from 'primereact/button';
 import { Chart as ChartJS } from 'chart.js';
+import { getCurrentDateTime } from '@/utils/dateUtil';
 
 const LineChart = () => {
     const [chartData, setChartData] = useState({});
@@ -198,19 +199,7 @@ const LineChart = () => {
 
         setChartData(data);
         setChartOptions(options);
-    }, []);
-
-    const getCurrentDateTime = (): string => {
-        const now = new Date();
-        const year = now.getFullYear();
-        const month = String(now.getMonth() + 1).padStart(2, '0'); // Months are zero-based
-        const day = String(now.getDate()).padStart(2, '0');
-        const hours = String(now.getHours()).padStart(2, '0');
-        const minutes = String(now.getMinutes()).padStart(2, '0');
-        const seconds = String(now.getSeconds()).padStart(2, '0');
-    
-        return `${year}-${month}-${day}-${hours}-${minutes}-${seconds}`;
-    };
+    }, []);    
 
     const handleExportToPDF = () => {
         const chartInstance = chartRef.current?.getChart(); // Access the Chart.js instance
