+++ styles/demo/Demos.scss
@@ -1,5 +1,1 @@
 @import './code.scss';
-@import './BlockViewer.scss';
-@import './TimelineDemo.scss';
-@import './flags/flags.css';
-@import './badges.scss';
